PP-14193 Move test with your users controllers to simplified account dir
