+++ src/controllers/simplified-account/services/test-with-your-users/index.controller.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { response } = require('../../utils/response.js')
+const { response } = require('@utils/response.js')
 const { formatSimplifiedAccountPathsFor } = require('@utils/simplified-account/format')
 const paths= require('@root/paths')
