+++ src/controllers/simplified-account/services/test-with-your-users/submit.controller.js
@@ -2,14 +2,14 @@
 
 const lodash = require('lodash')
 
-const logger = require('../../utils/logger')(__filename)
-const { response } = require('../../utils/response.js')
-const paths = require('../../paths')
-const productsClient = require('../../services/clients/products.client.js')
-const productTypes = require('../../utils/product-types')
-const publicAuthClient = require('../../services/clients/public-auth.client')
-const { isCurrency, isHttps, isAboveMaxAmount } = require('../../utils/validation/field-validation-checks')
-const { penceToPounds, safeConvertPoundsStringToPence } = require('../../utils/currency-formatter')
+const logger = require('@utils/logger')(__filename)
+const { response } = require('@utils/response.js')
+const paths = require('../../../../paths')
+const productsClient = require('@services/clients/products.client.js')
+const productTypes = require('@utils/product-types')
+const publicAuthClient = require('@services/clients/public-auth.client')
+const { isCurrency, isHttps, isAboveMaxAmount } = require('@utils/validation/field-validation-checks')
+const { penceToPounds, safeConvertPoundsStringToPence } = require('@utils/currency-formatter')
 const { formatSimplifiedAccountPathsFor } = require('@utils/simplified-account/format')
 
 module.exports = async (req, res) => {
