+++ src/routes.js
@@ -15,7 +15,6 @@ const { NotFoundError } = require('./errors')
 const { enforceUserFirstFactor, redirectLoggedInUser } = require('./services/auth.service')
 const trimUsername = require('./middleware/trim-username')
 const permission = require('./middleware/permission')
-const restrictToSandboxOrStripeTestAccount = require('./middleware/restrict-to-sandbox-or-stripe-test-account')
 const inviteCookieIsPresent = require('./middleware/invite-cookie-is-present')
 
 // Controllers
@@ -33,7 +32,6 @@ const registerController = require('./controllers/subscribe-service.controller')
 const createServiceController = require('./controllers/create-service/create-service.controller')
 const selectOrgTypeController = require('./controllers/create-service/select-organisation-type/select-organisation-type.controller')
 const inviteValidationController = require('./controllers/invite-validation.controller')
-const testWithYourUsersController = require('./controllers/test-with-your-users')
 const paymentLinksController = require('./controllers/payment-links')
 const twoFactorAuthController = require('./controllers/user/two-factor-auth')
 const feedbackController = require('./controllers/feedback')
